relocate code to correct files
