+++ packages/web/app/.env.template
@@ -35,6 +35,9 @@ AUTH_LEGACY_AUTH0_AUDIENCE="<sync>"
 AUTH_LEGACY_AUTH0_INTERNAL_API_ENDPOINT=http://localhost:4000/__legacy
 AUTH_LEGACY_AUTH0_INTERNAL_API_KEY="<sync>"
 
+## Organization level Open ID Connect Authentication
+AUTH_ORGANIZATION_OIDC=0
+
 # Slack
 
 INTEGRATION_SLACK="0"
