+++ packages/web/app/package.json
@@ -79,8 +79,8 @@
     "valtio": "1.2.5",
     "wonka": "4.0.15",
     "yup": "0.32.11",
-    "supertokens-auth-react": "0.26.3",
-    "supertokens-node": "12.0.2",
+    "supertokens-auth-react": "0.27.0",
+    "supertokens-node": "12.0.5",
     "supertokens-js-override": "0.0.4",
     "nextjs-cors": "2.1.1",
     "@whatwg-node/fetch": "0.4.7",
